// src/types/Producto.tsx
export interface Producto {
    idproducto: number;
    nombre: string;
    precio: number;
    preciorebajado: number;
    imagenchica: string;
    imagengrande: string;
    unidadesenexistencia: number;
    detalle: string;
    descripcion: string;
    categoria: string;
    proveedor: string;
    pais: string;
    telefono: string;
}